renamed schema, updated readme
