+++ src/routes/swagger.json
@@ -1247,7 +1247,7 @@
 	},
 	"info": {
 		"title": "@digicatapult/veritable-cloudagent",
-		"version": "0.4.1",
+		"version": "0.4.0",
 		"description": "Rest endpoint wrapper for using your agent over HTTP",
 		"license": {
 			"name": "Apache-2.0"
